thermal: sun8i: add R329 THS
The thermal sensor controller on R329 is similar to the one on H6, but
with only one sensor. Calibration method is still unknown because no
calibration is implemented in BSP kernel, neither is it documented in
the user manual.

Add its support to sun8i-thermal driver. Calibrating it is now
prevented.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>